+++ mapper/lib/src/model/annotations.dart
@@ -178,20 +178,20 @@ class JsonProperty {
       (jsonProperty.notNull == true || jsonProperty.notNullMessage != null);
 
   static String getPrimaryName(JsonProperty jsonProperty) =>
-      jsonProperty != null &&
-              jsonProperty.name is Iterable &&
-              jsonProperty.name.isNotEmpty
+      jsonProperty != null
+          ? jsonProperty.name is Iterable && jsonProperty.name.isNotEmpty
           ? jsonProperty.name.first
-          : jsonProperty.name;
+          : jsonProperty.name
+          : null;
 
   static List<String> getAliases(JsonProperty jsonProperty) =>
       jsonProperty != null &&
-              jsonProperty.name is Iterable &&
-              jsonProperty.name.length > 1
+          jsonProperty.name is Iterable &&
+          jsonProperty.name.length > 1
           ? jsonProperty.name
-              .where((x) => x != getPrimaryName(jsonProperty))
-              .toList()
-              .cast<String>()
+          .where((x) => x != getPrimaryName(jsonProperty))
+          .toList()
+          .cast<String>()
           : [];
 
   @override